requirements.txt
requests==2.28.2
pyyaml==6.0
prometheus-client==0.16.0
aiohttp==3.8.4
python-telegram-bot==20.2
pytest==7.3.1
pytest-asyncio==0.21.0
